+++ public/assets/js/selector.js
@@ -4,3 +4,10 @@ const terminalWrapper = document.querySelector('.terminal_wrapper');
 const terminalContainer = document.querySelector('.terminal_container');
 const usernameInput = document.querySelector('.username_input');
 const roomIdInput = document.querySelector('.room_input');
+
+
+var user;
+var localUser;
+
+var audio1 = new Audio('./assets/sounds/audio_1.mp3');
+var audio2 = new Audio('./assets/sounds/audio_2.mp3');