+++ CHANGELOG.md
@@ -1,9 +1,22 @@
+#### V 1.2 - Message Sound Notification
+
+Sat, 10th April 2022 9:00 AM
+
+- **Commands**
+  - /sound
+- **Uppercase command compatible**
+- **Non command message compatible Eg. /hello as a message and not command**
+
+---
+
 #### V 1.1 - Name Command Added
 
 Sat, 9th April 2022 7:50 PM
 
 - **Commands**
   - /name
+
+---
 
 #### V 1.0 - Initial Build